Try to have dynamic PageLayout
We have 3 roles with different sidebar . we need to have 3 layouts but, with some changes we can use the PageLayout component for difference roles such as admin, freelancer or owner.
I started with OwnerPageLayout component then, step by step I completed it.
Because the changes I have to fix some responsive approach.