Use radio button to handle filters for small screen
Instead of dropdown now, we have radio button in small screen to handle filters.
Write a hook that detect size of screen, so in small screen we can ignore some elements.
Correct some errors in filter dropdown.
Because I want to handle set params when a button clicked, so I use state.